requirements.txt
psycopg2-binary
requests
python-dotenv
streamlit-folium
sqlalchemy
streamlit-card
googlemaps
numpy
datetime